+++ CHANGELOG.md
@@ -8,9 +8,11 @@ BREAKING CHANGES:
   - MQ service server must be run together with its main server pair.
 
 IMPROVEMENTS:
+
 - Check ial at API when IDP response
 
 BUG FIXES:
+
 - Fix check validity of secret at create response (if invalid, the response will not store in blockchain).
 
 ## 0.10.1 (October 9, 2018)