+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## TBD
 
 IMPROVEMENTS:
+
 - Check ial at API when IDP response
 
 BUG FIXES:
+
 - Fix check validity of secret at create response (if invalid, the response will not store in blockchain).
 - Fix wrong callback result for create identity.