+++ CHANGELOG.md
@@ -14,6 +14,7 @@ IMPROVEMENTS:
 BUG FIXES:
 
 - Fix check validity of secret at create response (if invalid, the response will not store in blockchain).
+- Fix wrong callback result for create identity.
 
 ## 0.10.1 (October 9, 2018)