+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 BREAKING CHANGES:
 
-- Separate MQ module from main server process into another process to provide a more flexible deployment strategy.
+- Separate MQ module from main server process into another process to provide more flexible deployment strategies.
   - MQ service server must be run together with its main server pair.
 
 IMPROVEMENTS: